Autherized user access for rest verbs with JWT tokens

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException} from '@nestjs/common';
+import { Injectable, Logger, NotFoundException} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -10,6 +10,7 @@ import { TaskRepository } from './task.repository';
 
 @Injectable()
 export class TasksService {
+private logger = new Logger('TasksService');
 
 constructor(
   @InjectRepository(TaskRepository)
@@ -21,8 +22,8 @@ constructor(
     ):Promise<Task[]>{
   return this.taskRepository.getTasks(filterDto , user);
   }
- async getTaskById(id : number) : Promise<Task>{
-    const found = await this.taskRepository.findOne(id);
+ async getTaskById(id : number, user: User) : Promise<Task>{
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id} });
     if(!found){
     throw new NotFoundException(`Task with id "${id}" not found` );
     }
@@ -35,19 +36,19 @@ constructor(
     return this.taskRepository.createTask(createTaskDto , user);
   }
 
-  async deleteTask(id : number) : Promise<void>{       
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id : number, user : User) : Promise<void>{       
+    const result = await this.taskRepository.delete({ id, userId: user.id}); // delete tasks belongs to the current user
     if(result.affected === 0){
       throw new NotFoundException(`Task with ID : "${id}" not found` );
     }  
-    console.log(result);
-  }  
+   // console.log(result);
+  }   
 
-  async updateTaskStatus(id : number , status : TaskStatus):Promise<Task>{   //this is a single task
-    let taskstate = await this.getTaskById(id);
-    taskstate.status = status;
-    await taskstate.save();    
-    return taskstate; 
+  async updateTaskStatus(id : number , status : TaskStatus , user : User):Promise<Task>{   //this is a single task
+   let taskstate = await this.getTaskById(id, user);
+   taskstate.status = status;
+   await taskstate.save();    
+    return ; 
   }
 
 }
